refactor: Create redskyctl/util package
Moving some more generic/higher level functions into the util package for reusability.
